refactor: move sync blocks to SyncBlocksService
